Improved Deletion & Bug Fix

+++ src/editor/view.rs
@@ -51,8 +51,11 @@ impl View {
 
     // region: Text Editing
     fn backspace(&mut self) {
-        self.move_left();
-        self.delete();
+        if self.text_location.line_index != 0 || self.text_location.grapheme_index != 0 {
+            self.move_text_location(&Direction::Left);
+            self.delete();
+        }
+        
     }
 
     fn delete(&mut self) {
@@ -77,7 +80,7 @@ impl View {
         let grapheme_delta = new_len.saturating_add(old_len);
         if grapheme_delta > 0 {
             // move right for an added grapheme (should be the regular case)
-            self.move_right();
+            self.move_text_location(&Direction::Left);
         }
         self.needs_redraw = true;
     }
